+++ CSC8501-Coursework-Main-Assignment/ALLD.h
@@ -6,6 +6,6 @@ namespace ipd {
     public:
         std::string name() const override;
         Move nextMove(const MatchState& state, int selfIndex, Random& rng) override;
-        double complexity() const override;
+        int complexity() const override;
     };
 }
