+++ CSC8501-Coursework-Main-Assignment/Empath.h
@@ -12,7 +12,7 @@ namespace ipd {
         std::string name() const override { return "Empath"; }
         Move nextMove(const MatchState& state, int selfIndex, Random& rng) override;
         void reset() override;
-		double complexity() const override { return 3.0; }
+		int complexity() const override { return 3; }
 
     private:
         int m_remorseRemaining;
